Fix the libstdc++ and RGB-PNG bugs the Linux CI jobs exposed.
macOS and Windows now pass the whole pipeline; the earlier Windows failure was a
transient 504 fetching ninja, and the vcpkg port list resolved correctly. What
remained was Linux, and it found real bugs.

Thirty-seven headers used std::fabs, sqrt, floor, ceil, cos, sort, vector or
logic_error without including <cmath>, <algorithm>, <vector> or <stdexcept>.
libc++ leaks those transitively, libstdc++ does not, so the existing sweep could
not see them: it only ever ran against libc++. Homebrew's GCC provides libstdc++,
so the sweep now runs under both and the two agree exactly, 892 of 898 headers
self-contained either way. The include fixer understands GCC's phrasings as well
as Clang's now.

Five more, of which three are real defects rather than missing includes:

  * key_events.h declares enum class KeyCode : uint16_t without <cstdint>.
  * math_covariance.h called outer_prod<matrix3x3_type,vector3_type>(a,b), but
    outer_prod takes a single template parameter deduced from its arguments, so
    that call could never have compiled once instantiated. GCC checks template
    bodies eagerly and caught it; Clang defers and did not.
  * image_get_max_octaves.h called max(1u, i-2u) and grid_fast_blur.h called
    min(int, size_t); std::max and std::min need both arguments to be the same
    type, and unsigned int is not size_t on a 64-bit Linux.

The PNG test then segfaulted on Linux. AddressSanitizer localises it to a
1 KB heap overflow in image_read.h: png_get_channels() was queried before the
transformations were configured, so it reported the channel count of the file
rather than of the decoded rows. png_set_filler adds an alpha channel to RGB
input, so an RGB image allocated three channels while libpng wrote four. The
count is now read after png_read_update_info, where it is correct. This is a
pre-existing bug in the RGB path that only appeared because the committed test
fixture is RGB; the Wikimedia image it replaced was RGBA, where the two agree.

+++ OpenTissue/core/containers/grid/util/grid_fast_blur.h
@@ -9,6 +9,8 @@
 //
 #include <OpenTissue/configuration.h>
 
+#include <cmath>
+
 namespace OpenTissue
 {
   namespace grid
@@ -105,9 +107,9 @@ namespace OpenTissue
           im = ( i ) ?  im : 0;
           jm = ( j ) ?  jm : 0;
           km = ( k ) ?  km : 0;
-          ip = min( ip, I - 1 );
-          jp = min( jp, J - 1 );
-          kp = min( kp, K - 1 );
+          ip = min( ip, static_cast<int>(I) - 1 );
+          jp = min( jp, static_cast<int>(J) - 1 );
+          kp = min( kp, static_cast<int>(K) - 1 );
 
           //size_t idx   = ( k  * J + j )  * I + i;
           size_t idx_im  = ( k  * J + j  ) * I + im;
